Avoid running more unit tests in e2e target
