+++ tests/common.go
@@ -8,7 +8,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"math"
 	"os"
 	"os/exec"
@@ -22,10 +21,10 @@ import (
 	"github.com/stackrox/rox/pkg/pointers"
 	"github.com/stackrox/rox/pkg/protocompat"
 	"github.com/stackrox/rox/pkg/retry"
-	"github.com/stackrox/rox/pkg/retryablehttp"
 	"github.com/stackrox/rox/pkg/search"
 	"github.com/stackrox/rox/pkg/testutils"
 	"github.com/stackrox/rox/pkg/testutils/centralgrpc"
+	"github.com/stackrox/rox/tests/vmhelpers"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -61,25 +60,6 @@ const (
 var (
 	sensorPodLabels = map[string]string{"app": "sensor"}
 )
-
-// testutilsLogger adapts testutils.T to retryablehttp.Logger interface.
-//
-// WHY THIS EXISTS (and why it's unfortunate):
-// The retryablehttp library requires a logger implementing its Logger interface with Printf method.
-// Go's *testing.T has Logf but NOT Printf (different method names, same functionality).
-// We cannot add Printf to testutils.T because that would break compatibility with *testing.T,
-// which is used throughout the codebase (e.g., centralgrpc.GRPCConnectionToCentral(t testutils.T)).
-//
-// CLEANER ALTERNATIVE:
-// Accept *testing.T directly in helper functions and create testutils.T wrappers locally where needed
-// (specifically for the retry mechanism). This would avoid the interface constraint propagating everywhere.
-// However, this would be a larger refactor affecting many test helper functions.
-//
-// CURRENT COMPROMISE:
-// Use this tiny adapter ONLY where retryablehttp requires it. Everywhere else uses testutils.T naturally.
-type testutilsLogger struct{ testutils.T }
-
-func (l testutilsLogger) Printf(format string, v ...interface{}) { l.Logf(format, v...) }
 
 // logf logs using the testing logger, prefixing a high-resolution timestamp.
 // Using testing.T.Logf means that the output is hidden unless the test fails or verbose logging is enabled with -v.
@@ -730,12 +710,7 @@ func getConfig(t testutils.T) *rest.Config {
 // configureRetryableTransport configures a rest.Config to use retryable HTTP client
 // for network resilience. This adds automatic retry logic for transient network errors.
 func configureRetryableTransport(t testutils.T, restCfg *rest.Config) {
-	if restCfg.Timeout == 0 {
-		restCfg.Timeout = 30 * time.Second
-	}
-	retryablehttp.ConfigureRESTConfig(restCfg,
-		retryablehttp.WithLogger(&testutilsLogger{t}),
-	)
+	vmhelpers.ConfigureRetryableTransport(t, restCfg)
 }
 
 func createK8sClient(t testutils.T) kubernetes.Interface {
@@ -743,10 +718,7 @@ func createK8sClient(t testutils.T) kubernetes.Interface {
 }
 
 func createK8sClientWithConfig(t testutils.T, restCfg *rest.Config) kubernetes.Interface {
-	k8sClient, err := kubernetes.NewForConfig(restCfg)
-	require.NoError(t, err, "creating Kubernetes client from REST config")
-
-	return k8sClient
+	return vmhelpers.CreateK8sClientWithConfig(t, restCfg)
 }
 
 // getClusterID returns the ID of the cluster in Central.
@@ -894,10 +866,7 @@ func (ks *KubernetesSuite) logf(format string, args ...any) {
 	logf(ks.T(), format, args...)
 }
 
-type logMatcher interface {
-	Match(reader io.ReadSeeker) (bool, error)
-	fmt.Stringer
-}
+type logMatcher = vmhelpers.LogMatcher
 
 // waitUntilLog waits until ctx expires or logs of container in all pods matching podLabels satisfy all logMatchers.
 func (ks *KubernetesSuite) waitUntilLog(ctx context.Context, namespace string, podLabels map[string]string, container string, description string, logMatchers ...logMatcher) {
